Support text display in HFinder
HFinder creates local RastPort and BitMap (of depth 1) and does RectFill,
Move, Text and BltTemplate on these objects eventually copying (?) them
into ARGB buffer and using WritePixelArray to output this to window.

It might be possible to call 32-bit functions in this case instead of
current approach. BitMap created by HFinder is "Amiga" bitmap, not
HIDD bitmap, so it's possible that 32-bit code will work.